refactor: Split ast_tests.rs into focused test modules
Split the 1048-line ast_tests.rs file into 6 smaller, focused test modules
for better organization and maintainability:

- statements.rs (4 tests): DML statement tests (SELECT, INSERT, UPDATE, DELETE)
- expressions.rs (10 tests): Basic expression and literal tests
- operators.rs (13 tests): Binary and unary operator tests
- complex_expressions.rs (5 tests): CASE, subquery, and IN expression tests
- ddl.rs (2 tests): CREATE TABLE and DDL tests
- select_stmt.rs (19 tests): SELECT statement features (DISTINCT, GROUP BY, HAVING, LIMIT, ORDER BY, JOIN, subqueries)

All 53 tests continue to pass. This organization improves:
- Test discoverability by grouping related tests
- Navigation speed when searching for specific test cases
- Potential for parallel test execution
- Overall code maintainability

Closes #435
